requirements_qr_gov.txt
# Générateur QR Gouvernemental - Dépendances
# Core QR generation
qrcode[pil]==7.4.2
Pillow==10.1.0

# Performance & Parallelism  
numpy==1.24.3
multiprocessing-pool==0.1.1
concurrent-futures==3.1.1

# Database & Storage
sqlite3  # Built-in Python
redis==5.0.1
pymongo==4.6.0  # Alternative NoSQL
sqlalchemy==2.0.23  # ORM avancé

# System monitoring
psutil==5.9.6

# Web framework (API)
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0

# Message Queue
celery==5.3.4
rabbitmq==0.2.0
kombu==5.3.4

# Cloud Storage
boto3==1.34.0  # AWS S3
azure-storage-blob==12.19.0  # Azure
google-cloud-storage==2.10.0  # GCP

# Monitoring & Logging
prometheus-client==0.19.0
grafana-client==3.5.0
structlog==23.2.0

# Security
cryptography==41.0.8
passlib==1.7.4
python-jose==3.3.0

# Testing & Development
pytest==7.4.3
pytest-benchmark==4.0.0
locust==2.17.0  # Load testing
